requirements.txt
streamlit==1.32.0
pymongo==4.6.1
python-dotenv==1.0.1
pandas==2.2.0
numpy==1.24.3
pytest==8.0.2
python-dateutil==2.8.2
twilio==8.12.0
pydantic==2.6.1
bcrypt==4.1.2
PyJWT==2.8.0
email-validator==2.1.0.post1
python-magic==0.4.27
plotly==5.18.0
scipy==1.12.0
statsmodels==0.14.1
scikit-learn==1.4.0
matplotlib==3.8.2
seaborn==0.13.1 